It works!
Looks ahead 11 steps and chooses the step that leads to the best eventual board.
If the best board is available via multiple routes, takes the shortest route.

See demo.exs for a board that can be solved with lookahead of 11, but not with
a smaller number, because the initial board is on a local maximum requiring 11 moves
to improve.  Solving takes 4G and about a minute.

The AI currently gives up if it gets stuck in a local maximum.  Better would be to at
least move at random.

The program does not cache any values, and stupidly recomputes the full lookahead tree
after every move.  Caching would speed it up by orders of magnitude.